new update
removed location permissions
stream builder continuously checks if stream is active
file naming convention fixed
removed extensions
