+++ pubspec.yaml
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.3.0-244.0.dev <4.0.0'
+  sdk: '>=3.1.3 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -36,7 +36,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
   web_socket_channel: ^2.4.1
-  network_info_plus: ^4.1.0
+  # network_info_plus: ^4.1.0
   lottie: ^2.7.0
   easy_loading_button: ^0.3.2
   page_transition: ^2.1.0
@@ -45,13 +45,14 @@ dependencies:
   path_provider: ^2.1.1
   loading_animation_widget: ^1.2.0+4
   native_shutter_sound: ^0.0.1
-  social_share: ^2.3.1
-  permission_handler: ^11.1.0
+  # permission_handler: ^11.1.0
   image_picker: ^1.0.6
   photo_view: ^0.14.0
   image_cropper: ^5.0.1
   http: ^1.1.2
   image: ^4.1.3
+  share_plus: ^7.2.2
+  open_settings: ^2.0.2
